src/main/java/den/vor/easy/object/value/impl/BooleanValue.java
package den.vor.easy.object.value.impl;

import den.vor.easy.object.value.ScalarValue;

public class BooleanValue extends ScalarValue<Boolean> {

    public static final BooleanValue TRUE = new BooleanValue(true);
    public static final BooleanValue FALSE = new BooleanValue(false);

    public static BooleanValue of(boolean value) {
        return value ? TRUE : FALSE;
    }

    private final boolean value;

    public BooleanValue(boolean value) {
        this.value = value;
    }

    @Override
    public Boolean getValue() {
        return value;
    }

    @Override
    public String toString() {
        return "BooleanValue{" +
                "value=" + value +
                '}';
    }
}
